refactor: move `i18n.js` to `/lib/utils` etc
