+++ src/routes/+layout.svelte
@@ -1,22 +1,21 @@
 <script>
 import { navigating, page } from '$app/stores'
-import '@/app.css?v=280520231' // CSS file that is generated by TailwindCSS, version number is for caching purposes (so that the browser knows to download the new CSS file)
+import '@/app.css?v=30052023'
 import Footer from '@/components/Footer.svelte'
 import Nav from '@/components/Nav.svelte'
-import { startClient } from '@/i18n.js'
+import { startClient } from '@/lib/utils/i18n.js'
 import NProgress from 'nprogress'
 import 'nprogress/nprogress.css'
 import { derived } from 'svelte/store'
 
 NProgress.configure({
-    // Full list: https://github.com/rstacruz/nprogress#configuration
+    // https://github.com/rstacruz/nprogress#configuration
     minimum: 0.16,
     showSpinner: false,
-    // 3476CF
 })
 
+// nprogress bar
 $: {
-    // This is a reactive statement that runs whenever the $navigating store changes value (i.e. when the user navigates to a new page)
     if ($navigating) {
         NProgress.start()
     }
@@ -25,15 +24,16 @@ $: {
     }
 }
 
+// derived store that runs whenever the $navigating store changes value
 const delayedPreloading = derived(navigating, (_, set) => {
-    // this is a derived store that runs whenever the $navigating store changes value
     set(true)
     setTimeout(() => set(true), 1000)
 })
 
 startClient()
 
-$: segment = $page.url.pathname.substring(1).split('/')[0] // Get the first segment of the URL (e.g. /blog/1 => blog)
+// 
+$: segment = $page.url.pathname.substring(1).split('/')[0] // gets the first segment of the URL (e.g. /blog/1 => blog)
 </script>
 
 <Nav />
